Remove tryConstructTarget into barrel, move constructors to their files.
