+++ packages/cursorless-engine/src/processTargets/targets/LineTarget.ts
@@ -1,9 +1,10 @@
 import { Position, Range, TextEditor } from "@cursorless/common";
-import { BaseTarget, CommonTargetParameters } from ".";
+import { BaseTarget, CommonTargetParameters } from "./BaseTarget";
 import { Target } from "../../typings/target.types";
 import { expandToFullLine } from "../../util/rangeUtils";
-import { tryConstructPlainTarget } from "../../util/tryConstructTarget";
+import {tryConstructPlainTarget} from "./PlainTarget";
 import { createContinuousLineRange } from "./util/createContinuousRange";
+import {tryConstructTarget} from "./util/tryConstructTarget";
 
 export class LineTarget extends BaseTarget<CommonTargetParameters> {
   type = "LineTarget";
@@ -86,4 +87,21 @@ function getTrailingDelimiterRange(editor: TextEditor, range: Range) {
   return end.line + 1 < editor.document.lineCount
     ? new Range(range.end, new Position(end.line + 1, 0))
     : undefined;
+}/**
+ * Constructs a {@link LineTarget} from the given range, or returns undefined
+ * if the range is undefined
+ * @param editor The editor containing the range
+ * @param range The range to convert into a target
+ * @param isReversed Whether the rain should be backward
+ * @returns A new {@link LineTarget} constructed from the given range, or null
+ * if the range is undefined
+ */
+
+export function constructLineTarget(
+  editor: TextEditor,
+  range: Range | undefined,
+  isReversed: boolean
+): LineTarget | undefined {
+  return tryConstructTarget(LineTarget, editor, range, isReversed);
 }
+
