+++ packages/cursorless-engine/src/processTargets/targets/PlainTarget.ts
@@ -2,6 +2,7 @@ import {Range, TextEditor} from "@cursorless/common";
 import {Target} from "../../typings/target.types";
 import {BaseTarget, CommonTargetParameters } from "./BaseTarget";
 import {expandToFullLine} from "../../util/rangeUtils";
+import {tryConstructTarget} from "./util/tryConstructTarget";
 
 interface PlainTargetParameters extends CommonTargetParameters {
   readonly isToken?: boolean;
@@ -167,4 +168,21 @@ function getTrailingCharacter(editor: TextEditor, contentRange: Range): string {
   return end.isBefore(line.range.end)
     ? editor.document.getText(new Range(end.translate(undefined, 1), end))
     : "";
+}/**
+ * Constructs a {@link PlainTarget} from the given range, or returns undefined
+ * if the range is undefined
+ * @param editor The editor containing the range
+ * @param range The range to convert into a target
+ * @param isReversed Whether the rain should be backward
+ * @returns A new {@link PlainTarget} constructed from the given range, or null
+ * if the range is undefined
+ */
+
+export function tryConstructPlainTarget(
+  editor: TextEditor,
+  range: Range | undefined,
+  isReversed: boolean
+): PlainTarget | undefined {
+  return tryConstructTarget(PlainTarget, editor, range, isReversed);
 }
+
